- removed regions plugin directory
